// SGI_FRONTEND/src/models/GetDataUserModel/GetDataUserModel.ts
export interface GetDataUserModel {
        CT_CEDULA: string;
        CT_NOMBRE: string;
        CT_APELLIDO_UNO: string;
        CT_APELLIDO_DOS: string;
        CT_CORREO: string;
        CN_DEPARTAMENTO: number;
        CB_ESTADO: boolean;
        CT_CONTRASENA: string;
        CT_PUESTO: string;
        CN_TELEFONO: number;
        DESCRIPCION_DEPARTAMENTO: string;
        ROLES: number[];
}